feat: Move branding settings to a dedicated page
