+++ src/components/layout/AppSidebar.tsx
@@ -21,6 +21,7 @@ import {
   Crown,
   Lock,
   Sparkles,
+  Palette,
   User,
   HelpCircle,
   ArrowLeftRight,
@@ -278,6 +279,12 @@ const menuItems: MenuItem[] = [
         title: "General Settings",
         url: "/settings",
         icon: Settings,
+        feature: "settings",
+      },
+      {
+        title: "Branding",
+        url: "/settings/branding",
+        icon: Palette,
         feature: "settings",
       },
     ],
